Finish building out postgres auth
